Added indexing for complex orbitals/integrals
We lose one degree of symmetry from the 4-index integrals,
doubling the size of the UMAT array required to store the integrals.
This doesn't account for some integrals with repeated I,J or K,L
indices, and so will waste a little memory (decreasing with basis
set size) for a simpler indexing algorithm.

Call UMatConj(I,J,K,L,val) to take the complex conjugate of val if
and only if required by I,J,K,L to use the value stored at UMatInd