New logging option: **MULTTRUNCROFCIDUMP** [NoDumpTruncs] [NoTruncOrbs(1)]...[NoTruncOrbs(NoDumpTruncs)]
Default OFF.
This refers to the MP2 and CI natural orbital calculations.
It allows us to perform one calculation but print out multiple ROFCIDUMP files at the various truncation
levels.
The first arguement to follow the keyword is the number of ROFCIDUMP files to be printed.
This is followed by that number of integers which refer to the number of *SPIN* orbitals to be removed from
the virtual space.  These are put into an array named NoTruncOrbs(1:NoDumpTruncs).
Note, the old keyword **TRUNCROFCIDUMP** for just one truncation still works, it just sets NoDumpTruncs=1.