One more logging option added: **MULTTRUNCVALROFCIDUMP** [NoDumpTruncs] [TruncEvalues(1) ... TruncEvalues(NoDumpTruncs)]
This allows us to specify a cutoff for the eigenvalues by which to truncate the virtual orbitals in a MP2 or CI natural
orbital rotation.  The first arguement is an integer indicating the number of ROFCIDUMP files to be printed - this is
followed by this number of real arguements indicating the eigenvalue cutoff.  Orbitals with occupation numbers below this
will be removed and an ROFCIDUMP file printed.