Have changed the defaults slightly for open shell systems w.r.t brillouins theorem.
If an open shell system is detected, and brillouins theorem is being used (i.e we are not using
**NOBRILLOUINS** or **ROHF** or **ROTATEDORBS** and the singles are not being included in the
energy calculation) the calculation will crash and give an error message unless **UHF** is
explicitely used in the System block.  This means that even though it is an open shell system, we
still do not want to include single excitations in the energy.
This is because we more commonly use restricted orbitals and it is currently too easy to leave out
**NOBRILLOUINS** and get the wrong energy - it should be more difficult to do this now.