Remove some debugging info.
Also, fix bug in CALCMCSIZESPACE - orbitals cannot be picked twice.

SOFTEXIT facility also can be used in CALCMCSIZESPACE, where you can
create a softexit from a changevars file, in the same way as FCIMCPar
to exit a run. In order to get this to compile in serial, a serial version
of the changevariables routine has also been created in soft_exit.F90.

Also, the FciMCLogging module has been made to work & compile in serial and
parallel, which was also needed to get softexit to compile in serial.