**WRITETRANSFORMMAT** added to Logging block.
This keyword must be included if we are doing a natural orbital rotation and want to write out an MOTRANSFORM file.
This file is used with Qchem to get the cube files for the orbitals.
NOTE: This MOTRANSFORM file is currently only printed correctly when the code is compiled with PGI.

**USEHFORBS** added to System block to go with **ROTATEORBS**
This is particularly for debugging.  It allows us to use the HF orbitals to create ROFCIDUMP files or MOTRANSFORM files in the same way the rotated are fo
rmed.