Changed CASSTAR to **TRUNCINITIATOR**, and introduced **DELAYTRUNCINITIATOR** [IterTruncInit].
This is because we are talking about something a little bit more general than CASSTAR now, whereby we choose a space to be
the initiating space which can be CAS X Y or EXCITE X - both should work.

**DELAYTRUNCINITIATOR** is if we would first like to do a calculation in the truncated space and then allow spawning outside
this space with the initiator approximation.  This can be followed by an iteration at which inactive space spawning is allowed
or allows TRUNCINITIATOR in the CHANGEVARS file to initiate this.