**KEEPDOUBSPAWNS** option added.
This goes along with **TRUNCINITIATOR**.  It means that if two sets of walkers are spawned with the same sign on the
same determinant - these are allowed to live no matter where they came from.  This is different from before where
if these had both come from outside the active space they would have both been aborted.