New calc option **ADDTOINITIATOR** [InitiatorWalkNo] (this goes with **TRUNCINITIATOR**
This option means that if a determinant outside the initiator space becomes significantly populated, it is essentially
added to the initiator space and is allowed to spawn where it likes.
The minimum walker population for a determinant to be added to the initiator space is InitiatorWalkNo.

Have also added **SHIFTEQUILSTEPS** which determines the number of iterations after the shift is allowed to vary
that it begins to contribute to the average value in column 10.  The default is now 1000.

Also, printing out INITIATORStats which contains the statistics for the **TRUNCINITIATOR** method.

Yes I realise this should have been done in separate commits...