Split up if test.
Fortran does not require logical statements are evaluated
in any specific order (unlike C).  Thus for a stament like

t_arr .and. arr(1) == 0

can be evaluated in two ways:

* test t_arr is true and if so then test arr(1) is 0.
* test arr(1) is 0 and if so then test t_arr is true.

It is important that both parts are valid in order to avoid memory issues.
If this is not the case then the test must be written as a nested statement.