Fixed doubles excitation
The random and systematic generators now generate the same thing,
at least when starting from a determinant with nopen=0. Also got the
probabilities nearly right, but not quite.

Also fixed the singles so that no edge cases with repeated excitations.

Also added option to just change the Yamanouchi symbol to the gen_excits
routine.

Commited from the bus between Dover and Cambridge on the way back from
Tignes!