Bug fix to Arr and Brr if ENERGY-CUTOFF is used.
In many places it is assumed ARR(nbasis,2) and BRR(nbasis).  In actual fact
there are allocated using LEN (max # basis fns) rather than nbasis.  If
ENERGY-CUTOFF is used then LEN/=nbasis and problems ensue (e.g. the reordering
of Arr causes Arr(:,2) to be changed when it
shouldn't be.

Fix this by resizing Arr and Brr.

Ta to AJWT for help.