New option **WALKCONTGROW**.
This goes with the **READPOPS** option, and if present, it allows the number of walkers
in the calculation to continue growing until it reaches the number specified by
**INITWALKERS** in the input.  This is contrary to the default where the shift is allowed
to change as soon as the POPSFILE is read in.
This option is also turned on by default if the shift read in from the POPSFILE is exactly
0.D0.  This is based on the assumption that we would have likely started with the zero shift
and in the previous calculation, this hasn't started varying yet.