Knock knock. Who's there you say? A new keyword! **PARTIALLYFREEZEVIRT** [NoVirtOrbs] [NoElectrons]
This works in the same way as **PARTIALLYFREEZE** (which corresponds to the core orbitals) except
here the first integer corresponds to the number of virtual *spin* orbitals that will be frozen from
the highest energy down.  The second integer refers to the maximum number of electrons allowed to
occupy those 'partially frozen' orbitals in one determinant.  Spawning onto determinants with more
than this number of electrons in the frozen orbitals is forbidden.
This keyword goes in the INTEGRAL block with the standard **FREEZE** keyword, and can also be changed
dynamically.
[A whole lot of documentation is to come today by the way...].