Modified **MAXNOATHF** to add a second option arguement - [HFPopThresh].
If this is present, then if the HF population drops below MaxNoatHF(first arguement) - HFPopThresh,
then the shift is set back to 0.D0, and the number of walkers is allowed to grow again until
MaxNoatHF is reachieved.  If the second arguement is not specified, MaxNoatHF - HFPopThresh = 0, so the
population at the HF may drop to 0 with no consequences.