Tidying of soft_exit II.
ChangeVars was written twice, for parallel and non-parallel versions.

This made maintenance and extending ChangeVars annoying.

Combine the two versions into one.

This was largely a case of putting #ifdef PARALLEL ... #endif
statememts around MPI calls.

*However* there were some discrepancies between the serial and parallel
versions.  I merged them in a sane fashion, but this needs to be reviewed...