Move "library" source files into their own separate directory.
The idea is to keep things tidier and the main directory only for
neci-specific code.

This was motivated by the large number of header files that dSFMT
(the soon-to-be included RNG) brings with it.

We should clean up/merge util.F and util_mod.F90...

ps This causes problems with the dependencies for MurmurHash.  Do
rm -rf dest/depend to fix it.  Sorry!  It was necessary...