Faster Mersenne Twister implementation, dSFMT, added.
We fix the period of the random numbers to be 2^19937-1 (ie
identical to the original MT algorithm) in mkconfig.py.

dSFMT is written in C.

There is a wrapper, dSFMT_wrapper.cpp, in C++ which exposes
a (very limited) subset of dSFMT to Fortran.  It is easy to
expose more functionality as needed.

There is also an interface module which will store 5x10^4 random numbers and
generate more random numbers as needed.  This is much more efficient than
generating one random number at a time.