Remove SSE2 option from PC-PGI 32 bit config.
It seems compiling in 32 bit on 64 platforms is becoming increasingly
fragile, particularly when it involves accessing more esoteric libraries
(e.g. SSE2 operations).

We use PGI 8.0.5  in the nightly tests and this does compile with SSE2
on.  However, later versions do not compile the SSE2 optimisations.
Removing the SSE2 setting allows later versions to compile and (as
a bonus) stops the seg fault on the Li2_CCSD  test with 8.0.5: either
8.0.5 was linking to an incorrect library which miraculously compiled
(but then caused seg faulting) or this modifies the memory layout
sufficiently that the seg fault is avoided (for now).  Hopefully the
former especially as the CCSD test is the only serial test
which uses the Mersenne Twister...