First **FREEZEINITIATOR** FreezeInitIter option.
This sets the calculation to take the initiator space of a particular
calculation at FreezeInitIter, and freeze it, so that the initiator
determinants do not change from then on.

The problem at the moment is that determinants are removed from the
initiator space when its population falls to 0, and they cannnot be
recognised as being part of that instantaneous initiator space.
This needs to be fixed.