**HISTINITIATORPOPS** [MaxInitPop] [HistInitIter] logging option added.
This sets up an array from -MaxInitPop -> MaxInitPop, and histograms the
populations of the initiator determinants at every HistInitIter iteration.

InitPops-Iter files are then printed with the population followed by the
number of initiators with that population.

A warning is printed if an initiator exceeds the bounds of the array with
a population too large (this will happen with the HF pretty quickly, but
HF_pop is already being printed).