Option to put FLIPHIGHPOPSIGN in the CHANGEVARS file, and this will
flip the sign of the determinant with the largest -ve population (that is the
largest population with sign opposite to the HF).
I'm very much just playing around with this at the moment, will likely change
it a lot or get rid of it.

Also fixed the bug in the reading in of the binary POPSFILES