Use spin-orbital indices in UEG calculations if required.
If we're not using the full Coulomb potential for exchange integrals, then
we need to be able to detect exchange integrals.  When using spatial indices
an integral like < i i+1 | i+1 i > is mapped to < a a | a a >, making it impossible
to detect exchange integrals.

For instance

< 1a 2b | 2a 1b > /= < 1a 2a | 2a 1a >

when the full Coulomb potential is not being used to model exchange interactions.

James Shepherd worked round this by explicitly asking for exchange-type
integrals in the Slater-Condon rules.  This is a cleaner solution which
makes detection of exchange integrals easy by not discarding the
spin-information of the basis.