Just a few fixes here and there to the last part of the fitting.
I've left the Dv vs v fit in, even though it doesn't seem to really work.  I don't
think this is a problem with the code but with the very tiny values.
The paper doesn't seem to take any notice of this final fit and just uses the D_e values
from the bottom of the potential.