Set missing variable declarations.
A commit inspired by aa84ba.

George mentioned that a bug would have been much easier if implicit none
was given.  It occurred to me that it would be useful if we could force
compilers to impose an implicit statement (eg gfortran has the
-fimplicit-none flag).

Some true "highlights" (step forward Ali, David, Ramin, Alex(?)) of neci
don't use implicit none and so have variables which are implicitly typed.
This commit fixes this so the F90 files compile with gfortran -fimplicit-none.

.F files will probably be nastier, but it's a step in the right direction.
Feel free to chip in with dealing with the .F files. ;-)