Fix cascade compilation for use with pathscale.
Recompiling with pathscale is now substantially faster.

The problem was that pathscale produces module files with uppercase
filenames, yet the dependencies were generated assuming that the module
files had lowercase filenames.  This caused compile_mod.pl to check e.g.
module_name.mod rather than MODULE_NAME.mod and as module_name.mod of
course didn't exist, it forced re-compilation. So whilst the
dependencies were consistent, the modules that make expected to produce
from the dependencies were not actually being produced.

It is easy to fix this by having a set of dependencies for each case.
make depend produces both sets so that the dependencies are always consistent and the appropriate dependency file is conditionally included.

Am surprised that no-one who regularly compiles with pathscale (either on workstations or on HECToR) hadn't already sorted this out...