Starting to work on making a single walker bit representations combining all info. Much work to be done. Changes to far reasonably haphazard! Initially, all info to be added as seperate integers and compress later. However, although not initially implemented, we want it to also be as easy as possible to extend to complex walkers later.
Key changes which need to be implemented:

NIfTot has changed its meaning.
Routines which therefore want to look at a determinant and look at 0:NIfTot
now need to be changed. Bounds are going to have to be checked carefully.

All mention of CurrentSign wants to be removed. Signs are now an array of dimension
lenof_sign, so this has to be put in. At the moment, we only want to consider
element 1 of this array though.