Slight change to the way the initiator flags are used.
Now, in the currentdets array the initiator flag refers to whether on not
that particular determinant *itself* is an initiator.
In the spawneddets array however, the flag refers to whether or not the
determinants *parent* was an initiator.

This just makes things a bit more clear when making changes etc.

Also, a new keyword is now possible (in the calc block):
**RETESTINITPOP** [OFF]

This is on by default and corresponds to the original algorithm where the
populations of initiator dets are tested each time, and these can become
non-initiators again if their pops fall too low.
Using the OFF setting however sets the initiators to remain initiators
once they become initiators, regardless of their population.
(However, if their pop completely dies and they are removed from the list,
their initiator status is no longer saved).