Obtain correct conjugation of the integrals from UMAT.
This handily uses an existing function, so the changes to UMATIND
where not necessary.

Note that care *must* be taken when setting UMAT elements, as UMAT
might actually need to store the conjugate form of the integral.

Similarly access should always be via the getumatel functions so that
conjugation is taken care of.