Bug fix to dSFMT interface.
Correct reinitialisation of dSFMT.

Previously dSFMT was being reinitialised but the index of the next
pseudo-random number was not reset to 1, so the exact string of PRN were
not being obtained.

Results will change (not within statistical error) if calculations
involve restarting dSFMT---e.g. if you did CALCMCSPACESIZE and FCIQMC
calculations within one run of neci.