FCIMCStats files will no longer be overwritten.
If reading in a POPSFILE, and an FCIMCStats file is present,
it will append to the file.

If not reading in a POPSFILE, and an FCIMCStats file is present,
then it will rename the FCIMCStats file FCIMCStats.x, where
x is the largest free integer.

Therefore, FCIMCStats will always refer to the current calculation,
and the previous ones will be labeled FCIMCStats.x -> FCIMCStats.0,
with FCIMCStats.0 being the oldest run.

This convention is the same as the MOLPRO output files.