New walker_death routines - EXPLICITHFRDM not working.
The walker_death routines have been split so that we calculate the
number of walkers to die for a particular Di at the beginning of the
run over ni - but this is just stored (and used for the
rdmbias calculation) but not actually changed in currentdets until
the end like usual.

*This will change the random number seed sequence* and so will change
tests.