Removed EXPLICITHFRDM completely from the code.
It was confusing the code, and it is difficult (and costly) to impliment
with the new walker_death routines.
It also seemed to give a false impression of convergence of the RDM
energy.

It appeared to converge very quickly (as the singles and doubles are
well accounted for) but the stochastic contributions from the rest of the
space are still contributing gradually, and in fact crucial to get the
right energy.