Tau searching.
If TAU keyword is left out, then the optimal tau will be dynamically
searched for, until the simulation enters variable shift mode.

The optimal tau, is one which doesn't allow for spawning events
of more particles than n_add in one attempt, or alternatively,
5 walkers at a time in the full scheme.

Instead, the maximum number of spawning events per attempt can be
set in the input, using the keyword:

MAXWALKERBLOOM <Max number of walkers allowed in one attempt>

The simulation will not be restart, but proceed as normal if tau
is changed during the simulation.