New Logging option **RDMGHOSTCHILD** [GhostThresh] [GhostFac]
This means that if p_spawn is less than GhostThresh, it is multiplied
by GhostFac, and if a child would have been spawned with this new value,
a child is created with sign zero.  This means the DiDj pair in
question can contribute to the reduced density matrix.

The number of extra spawns this creates needs to be tested for
larger systems than H2, but it seems there are not too many extra.