H2O_MC bug fixed
This was extremely unpleasant to debug as large numbers of routines had been moved around
in a single commit.

Firstly, there were some uninitialized variables, and lacking nullifies.

Next was array bounds.  The gist of it was that an array (SymProds) in DSTORE in the symmetry excitation
generators was being passed as an array of length 1.  This was ok when declared as an assumed-size
array as it can't be copied, and is passed in its present position, and accessed only up to its
allocated (but not actually known at calling) length.
This had been later redeclared as an adjustable array using its maximum possible size (which it does
have during sym gen setup).  These two did not interact well, and no data was being returned in the array.
(a compiler copy perhaps?  Too horrific to contemplate)

Fixed by actually specifying the correct length of the DSTORE array, and specifying the correct length
subset of it to be passed as SymProds.

/home/ajthom/NECI/SOURCEn