Dump_Forces_Info should now be correct.
Extraneous write statements removed.

Note: this version still recalculates the RDM energy from the symmetrised
chemical notation RDM, to check this has been done correctly.  Once we have
a few successful force results from this work, this "debugging" output can be removed.

We probably still want to print something at the end of the run giving the error in the
'symmetry' of the Lagrangian matrix, rather than printing out the matrix itself.  This
will be analagous to the Hermiticity error calculation.  We also still want to symmetrise
(ie, average D_ij and D_ji) when we print out the 1RDM.