Bug fix in Annihilation.
"ToRemove" was being incremented incorrectly.

In a sitation where the SpawnedParts array contains a zero sign after
compression, this zero was being counted twice, and causes the code to
crash.  Counting was being done once at the time of compression, and
once in AnnihilateSpawnedParts.  This has been changed such that the
second count in AnnihilateSpawnedParts only contributes if the sign
after compression was non-zero, to prevent this double counting.