Bug fix for RDMs.
Deals with the case where the sign on a determinant is instantaneously zero in the middle of the
annihilation routine.  I.e., previously sign 1.  Walker dies.  Walker spawned onto same det in
same iteration.  Because of the way annihilation is arranged, the CurrentH array is reordered,
and has its zeros removed at this point during annihilation, so that the average sign info
stored in CurrentH is lost and rezeroed, even though the determinant instantly becomes reoccupied.

To match this, we have to rezero AvNoAtHF at this point as well.