Changed attempt_create_normal and walker_death routines to match those currently in master.
The revised versions of these routines in the standard realcoeff branch follow different
trajectories to the old versions in master, making it impossible to accurately test changes
in this branch against results from master.  This current branch has therefore been created
to match the trajectories followed by master, but including both real coeff information and
dmat calculation.  This means we can do "integer" runs on this branch, and check that the
density matrices, and FCIMCStats files (etc) are identical between master and this branch,
indicating that the merge of the realcoeff and dmat information has been done successfully.

Note that this branch still contains a major compile error in ifort, but has been tested successfully
running in serial with gfortran.  Updates to follow.