Don't apply occupation threshold to states in the core space
Small change to ensure the occupied thresh option (where we round the population of low-weight dets up/down to save memory/time) so that it's only applied for determinants which are not in the core space.  We always keep the core-space dets in currentdets, so there's nothing to be gained from discretising here.

Also added a stop_all to reflect the fact I haven't fixed things up to work for HPHF + RDM + SS. (Phew!  It's all so very complicated :-/)