Added some additional statistics to keep track of the size of the occupied real space etc.  These are just to help me diagnose exactly what's what, and should be removed again eventually, I should imagine.
Also a very important change to walker_death as the subroutine arguments were in the wrong order.  AGAIN! It seems I fixed it in the other branch back around xmas time, but forgot to fix it here.  :S  Looks like it's only about a weeks' worth of calcs I have to redo though, rather than months so that's an improvement I guess....