Changed some data types from double precision to real(dp) in molpro_dump_mcscf_dens_for_grad.
Previously we were passing REAL*8 types into the routine which demanded REAL*16.  Failed compilation on gfortran.  Changed the latter to REAL*8.

Presume they were originally taken directly from the declarations in the equivalent Molpro routine, so needed to be changed to match our own setup.  I believe we don't ever intend this fciqmc version of the routine to be called once we're interfaced with molpro anyway -- the molpro version would be called instead.  Although perhaps this will produce problems of its own, if molpro is expecting REAL*16 types for the density matrices, whilst we're passing in REAL*8.... Should be easy to fix if this turns out to be a problem though.