Series of changes to allow us to propagate two independent runs simultaneously.
This makes use of the complex code (lenof_sign=2) but has a new parameter (inum_runs) that indicates we're doing a double run.  To use this mode, you need to compile with -D__DOUBLERUN under cppflags. I'll add a config file soon.

There are quite a few features that haven't yet been plumbed in. As such, you can't use the double run code in conjunction with:
    -   semistochastic
    -   ccmc
    -   density matrices
    -   readpops (concerning!)
amongst others.  I've put in stop_alls where I know there's a problem, but it's likely that there are other things that I haven't yet considered that may or may not be ok.

I intend to fix it to work with readpops, though this may need a bit of thought as we now have 2 shifts.
I'm also intending to fix density matrices for the next commit.  SS should happen at some point soon as well, though I might need to delegate that one...