Changes to accumulate the RDMs correctly (I hope) whilst doing a double run.
This works by using one sign from each population when adding in a product of coefficients for the RDMs.  This is particularly important for the diagonal elements -- as the two c_i s are now strictly uncorrelated, we shouldn't have any systematic bias.

Bear in mind that we only take off-diagonal connections from successful spawns within set 1 at the moment - this is to make sure that all the unbiasing probabilities work out correctly. We use one sign from each population though, as normal.