We can now print out RDM energies at intermediate points without introducing any bias.
This does use more memory than before.

Also fixed an unfortunate bug in attempt_create_normal which has caused the RDMBiasFacCurr elements
to be incorrect when the magnitude acceptance probability is greater than 1, and has negative sign.
This error was also present in the realcoeffs branch.

Of course, this development has brought me great personal sadness as I now have to redo an unknown
quantity of work.  Based on previous form, I imagine it will be A Lot. @£&*%$!