Correction to recent changes for create_particle_with_hash_table
Some of the KP-FCIQMC tests were failing because I had removed a block
of code in FindResidualParticle that makes it return if the population
is zero. I thought this was fine because the if statement later won't
set the initiator flag if the populaton is zero. However, the initiator
flag was being set in create_particle for these zero-weighted spawns.
This is incorrect, so has been fixed, so that the return statement is
not needed anymore. Both methods with and without a hash table for the
spawned list pass the tests now.