Make new benchmarks for tests that were failing.
Some of the tests were failing - a single KP-FCIQMC test was failing
because of numerical error in adding the spawned particles after the
change to use a hash table for the spawned list. The initiator complex
tests were failing because the definition of the initiator method
changed a little, but I believe that the new defintion is as was
intended, so I have kept these new benchmarks instead.