Fix crazy bug, and some other minor changes
Fix bug whereby semi-stochastic contributions to the projected Hamiltonian
are inconistent with the way that other contributions were added in, which
was causing an issue with the recent change to allow us to cycle on the
Hartree-Fock, and to allow exact spawning, when used together.

Also, by default never attempt deterministic spawning, and use the new
initialisation routine for the deterministic spawning, and add
support for k-point symmetry for this deterministic spawning.