Start work on being able to use different trial states for different replicas.
This includes, cleaning up of the trial initialisation routine, using the
Lanczos wrapepr, which can return excited states instead of just the ground
state. To use this, I've also moved all parameters for creating subspaces
to the input derived type, so that create_trial_states can be reused.

I've also removed the low-energy option, which I think is pointless, and
removed the tests for this option with it.

Finally, the names 'CORESPACE' and 'TRIALSPACE' were simply changed
to 'DETFILE', again so that the same routine can be reused.