Move replica overlaps to the final data table columns, and add option to not print them.
It was getting a bit annoying to have to look up which columns to
print out every time you change the number of replicas. Now these
columns should be constant, because the overlaps (which were causing
the varying) are in the final columns.

Also, add the option to not print out the replica overlaps, in case
they are not wanted (they can take up a lot of memory).