Add code to find the optimal trial states for each FCIQMC wave function.
This commit adds a routine to the start of the trial WF initialisation
that calculates the overlaps between all trial wave functions and all
FCIQMC replicas, and assigns the best trial wave function to each
replica.

It also adds an input option to allow the user to ask more trial
excited states be calculted than need to be kept, which can be
important.

Not yet implemneted (and probably broken by this commit): using
pairs of replicas, as in KP-FCIQMC or RDM calculations.