Don't allow the entire population sign to be flipped by default.
This messes up the blocking analysis, because the numerator and
denominator separately undergo a massive change from positive to
negative or vice versa, which leads to a massive error in the
blocking analysis, or you have to leave miss out a lot of data.
Flipping the entire population sign should not be necessary, but
has been kept as an option.