Don't compare reals against 0.0_dp, in multiple places throughout code.
In many of the new tests in the test suite (seemingly only those using
mneci, the new excitation generator and tau search in parallel), the
ix86-gfortran64-MPI-theor configuration is leading to failing tests.
Tracking this down, it seems to be because of multiple places in the
code where signs have been checked against 0.0_dp with an exact
equality. Numerical errors means that differences occur between
different compilers and the tests fail.

I have gone throguh and fixed these if statements in many places, although
I don't think I have got them all, but the situation should at least
be improved.

There was even a need to change the initiator threshold to
init_thresh+1.e-12_dp, because the tau search makes it so that some
spawns aim to have exactly the initiator threshold.