Undo change to intitiator threshold, instead reduce tau when using tau search
In a recent commit (6289672b3) I added 1.e-12_dp to the initiator threshold
to ensure that spawns exactly equal to the initiator threshold (coming
from the tau search option) don't cause discrepencies across different
compilers due to numerical errors.

I have undo this change (there would have been multiple other places
to change this, to do it properly) and instead made it so that tau
is updated to be slightly smaller, when using the tau search option.