Add an alternative approach to orthogonalising when we have replica pairs.
Before, we attempted to use as unbiased an orthogonalisation term
as possible. However, in doing so, this introduces a correlation
between the pairs of replicas that lead to the RDM estimates.

So instead, introduce an orthogonalisation procedure that
orthogonalises all of the even numbered simulations together,
and all of the odd numbered simulations together, so that all
even and odd simulations will be uncorrelated.