Add RDM columns individually, rather than summing whole RDM arrays in one go
ifort creates a lot of temporary arrays on the stack. An example
of this seems to be in the lines changed in this commit. This causes
problems with large RDMs, where we get stack overflow errors.

To get around this, sum RDMs together column by column, rather than
in one go.