Compare to 0.0_dp exactly in RDM check, instead of some small number
Usually this is bad, but here it makes sense because it is
guarenteed that if the number is not exactly zero then it must be
the replica that we are looking for, bacsue on how this sign array
is set up.

Also, change a cycle statement to an exit statement, which it should
have been before.