Update test suite input files to turn off load balancing.
Load balancing has recently been made default. This is causing tests
to fail (and crash), so should be turned off.

Also, in a few instances remove the linscalfcimcalgo option, which
was made default (and indeed the only option) a while back, so this
is not needed in the input file.