Add pops-core-approx option, which uses less memory than pops-core for large processor counts
The pops-core option stores a list on each process whose size scales
linearly with both the number of processors and the deterministic
space size, causing problems when both of these are large.

In this commit we introduce an approximate (but usually very accurate)
alternative, which only looks at the best 10*M/N states on each
process, where M is the deterministic space size and N is the number
of processors. Thus, the combined list size is 10*M, and no longer
scales with the number of processors.