Remove stop_all in RDM code, regarding off HPHF calculations
There is a stop_all when a closed-shell det is found in an odd-S RDM
calculation. But this isn't always desired because sometimes that
function is called with the "true" HF, which is always closed shell,
even in odd-S calculations. This doesn't cause problems because the
HF sign is 0, so no contribution is added from this undesired
state, but the stop_all prevents the calculation from running.