2nd major GUGA commit:
sorry again simon.. shame on me!
in this commit, the finished version of the weighted and symmetry
adapted GUGA excitation generator and a new automated tau-search
implementation for the old, non-weighted, non symmetry adapted
GUGA excitation generator are implemented.
Also major bug fixes and optimization for both approaches are
included.