Finished intitialisation routines for a single GF
Finshed the implementation for a single GF element
for both lesser and greater greensfunction
This includes the preperation of the wavefunction
to project onto
This is until now only done for a single-run
so no use of multiple replicas!
Next step is to set up the new real-time FCIQMC
main loop with the changed walker dynamics
corresponding to the real-time Schroedinger equation!