Fix another potential bug in the RDM code, involving an optional parameter.
Code uses optional parameter in if statement - it should really use the
parameter which is always present, and false if the optional parameter is
not entered. It shouldn't have affected any previous results as it is
only for semi-stochastic + singlerun + optional diagonal cutoff, which
is never used together in practice.