Code to print out 2-RDMs with spin. Also, redefine the calculation of the processor label.
Routine added which prints out the 2-RDMs, as in TwoRDM files previously.
One difference currently is that Hemiticity is not made use of in the output.
The names of the files also needs changing later.

The rows of the RDMs were distributed across processors in the undesired
way - q was dominant over p, but we want it the other way around, so that
no communication is required to output these particular files. This
has been corrected now.