Output a final energy for each state being sampled
Now, in the final output, a final energy estimate is output
for each states sampled.

The interface to FciMCPar has been modified to allow this, and the
Weight parameter has been removed from this input, as it is
presumably very archaic and doesn't seem to be used now.

Also, the text accompanying the final energy has been changed
to something less odd. This will require changing the test suite
extract script to search for the new text string.