Worked to run excited states wigh GUGA
Started working to get the excited states calculation
running with the GUGAs.
Changed the projected energy list to deal with multiple
runs and adapted the proj. energy calculation in the
necessary NECI parts.
Still not quite sure if i found all the parts,
since in some calculation the projected energy
and the shift value for excited states do not
seem to agree yet

Also to note: I have done that really naively for now
and I am not using any options like DOUB_INIT
but only plain, single reference states

In the future, make the semistochastic finally working
with GUGA too, and then use the necessary codes also
to create trialwavefunction functionality... todo!