fixed projected energy calc. for multiple replicas
fixed a bug, where the wrong projected energy list was used
to calculate the energy for the excited states in the GUGA
calculation

Also changed the off_diag_ref matrix element routine to
additionally output the excitation level to the
current reference determinant to correctly handle the
excitation level up to double excitation also for GUGA
simulation. But still can only handle this identification
up to double excitations!