Finished full semi-stochastic implo. for GUGA
Finished most of the aspects of the semi-stochastic implementation
for the GUGA implementation. But tests still need to be done!!!

Done multiple methods, like most populated, optimised, mp1
singles and doubles etc.

In the process also fixed the sparse_hamilton calculation for
GUGA and other necessary remaining parts in the remaining
NECI code.