Perform communication of estimates in collective MPI calls, and communicate all components of all estimates
Now, communication of estimates and stats is done through two calls, one for
HElement_t(dp) quantities, and one for everything else (plus a few others for
MPI_MAX and MPI_MIN reduce operations).

Also, some of the arrays were only communicating the first component previously,
particularly for initiator stats. This now allows the entire arrays to be
communicated in a general way.